Vereinheitlichung der Package-Namen. Gleiche Namen unter rest, domain, services
