+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/wahlbezirke/WahlbezirkRepository.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlbezirk;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlbezirke;
 
 import java.time.LocalDate;
 import java.util.List;
