+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/wahlen/Wahl.java
@@ -1,8 +1,12 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlen;
 
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.time.LocalDate;
@@ -10,7 +14,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 import org.springframework.stereotype.Indexed;
 
 @Entity
@@ -19,24 +22,33 @@ import org.springframework.stereotype.Indexed;
 @EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "Wahltag")
-public class Wahltag {
+@Table(name = "Wahl")
+public class Wahl {
 
     @Id
     @NotNull
     @Size(max = 1024)
-    @ToString.Include
-    private String wahltagID;
+    private String wahlID;
+
+    @Size(max = 255)
+    private String name;
+
+    @Min(0)
+    private long reihenfolge;
+
+    @Min(0)
+    private long waehlerverzeichnisnummer;
 
     @NotNull
-    @ToString.Include
     private LocalDate wahltag;
 
-    @Size(max = 1024)
-    @ToString.Include
-    private String beschreibung;
+    @Enumerated(EnumType.STRING)
+    @NotNull
+    private Wahlart wahlart;
 
-    @Size(max = 1024)
-    @ToString.Include
+    @Embedded
+    private Farbe farbe;
+
+    @Size(max = 255)
     private String nummer;
 }
