+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/wahlen/Wahlart.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahl;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlen;
 
 public enum Wahlart {
     BAW, BEB, BTW, BZW, EUW, LTW, MBW, OBW, SRW, SVW, VE
