+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/wahltag/Wahltag.java
@@ -1,12 +1,8 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahl;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahltag;
 
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.time.LocalDate;
@@ -14,6 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 import org.springframework.stereotype.Indexed;
 
 @Entity
@@ -22,33 +19,24 @@ import org.springframework.stereotype.Indexed;
 @EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "Wahl")
-public class Wahl {
+@Table(name = "Wahltag")
+public class Wahltag {
 
     @Id
     @NotNull
     @Size(max = 1024)
-    private String wahlID;
-
-    @Size(max = 255)
-    private String name;
-
-    @Min(0)
-    private long reihenfolge;
-
-    @Min(0)
-    private long waehlerverzeichnisnummer;
+    @ToString.Include
+    private String wahltagID;
 
     @NotNull
+    @ToString.Include
     private LocalDate wahltag;
 
-    @Enumerated(EnumType.STRING)
-    @NotNull
-    private Wahlart wahlart;
+    @Size(max = 1024)
+    @ToString.Include
+    private String beschreibung;
 
-    @Embedded
-    private Farbe farbe;
-
-    @Size(max = 255)
+    @Size(max = 1024)
+    @ToString.Include
     private String nummer;
 }
