+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/wahltag/WahltagRepository.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahltag;
 
 import java.util.List;
 import java.util.Optional;
